exist directory

+++ install.sh
@@ -8,10 +8,20 @@ ln -s -f ~/.dotfiles/.zshrc ~/
 
 #emacs
 cd ~
-git clone https://github.com/uchiyu/.emacs.d.git
-ln -s -f ~/.emacs.d/init.el ~/
+dir=~/.emacs.d
+if [ -e $dir ]; then
+  git pull origin master
+else
+  git clone https://github.com/uchiyu/.emacs.d.git
+  ln -s -f ~/.emacs.d/init.el ~/
+fi
 
 #vim
 cd ~
-git clone https://github.com/uchiyu/.vim.git
-sh ~/.vim/install.sh
+dir=~/.emacs.d
+if [ -e $dir ]; then
+  git pull origin master
+else
+  git clone https://github.com/uchiyu/.vim.git
+  sh ~/.vim/install.sh
+fi
